Update default date range and rolling #
Originally I thought I would be must interested in my progress over
time. But I find I am actually most interested in my progress in the
last week. My process is:

1. Load up data on Monday
2. Change the dates to show from Monday of last week to Sunday so
   I am looking at my progress for the past week.
3. Change the rolling count to 1 since such a short range distorts
   rather than enhances.

By changing the defaults I can remove step 2 and 3. The feature to see
a larger time span and have rolling dates is still useful so keeping it
but just not the defaults.

In case there is not a weeks worth of data it does adjust to the
earliest date that we have data for.

Also has the advantage of loading the initial screen faster as it does
less processing (still parses and normalizes all data but doesn't do
the calculations for the entire range).