Improve task completion date for archived tasks in completed projects
If a project is marked completed with open tasks still then the task
is not considered completed. But when archived all those tasks suddenly
get completed on that archive date even though the project completion
date is probably a better date.

In this situation update the task completion date to be the project
completion date.