Sort tag charts by most first
This will make it easier to see what we are spending the most time on
(and the least). Kinda makes me wonder if this means I should bring the
0 totals back in since they are truly the least. But don't like the
clutter so leaving out for now.

`total` is added to the data sent to chart as a side-effect but the
extra data won't hurt anything and it will allow us to do our sorting.
This also gives us the benefit of being able to filter the 0 count items
much faster.